Module syntax_ext::deriving

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

The compiler code necessary to implement the #[derive] extensions.

Modules

bounds 	 [
Experimental
]

clone 	 [
Experimental
]

custom 	 [
Experimental
]

debug 	 [
Experimental
]

decodable 	 [
Experimental
]
The compiler code necessary for #[derive(Decodable)]. See encodable.rs for more.

default 	 [
Experimental
]

encodable 	 [
Experimental
]
The compiler code necessary to implement the #[derive(Encodable)] (and Decodable, in decodable.rs) extension. The idea here is that type-defining items may be tagged with #[derive(Encodable, Decodable)].

eq 	 [
Experimental
]

generic 	 [
Experimental
]
Some code that abstracts away much of the boilerplate of writing derive instances for traits. Among other things it manages getting access to the fields of the 4 different sorts of structs and enum variants, as well as creating the method and impl ast instances.

hash 	 [
Experimental
]

ord 	 [
Experimental
]

partial_eq 	 [
Experimental
]

partial_ord 	 [
Experimental
]

Functions

is_builtin_trait 	 [
Experimental
]

register_builtin_derives 	 [
Experimental
]