Struct syntax_ext::deriving::generic::FieldInfo

pub struct FieldInfo<'a> {
    pub span: Span,
    pub name: Option<Ident>,
    pub self_: P<Expr>,
    pub other: Vec<P<Expr>>,
    pub attrs: &'a [Attribute],
}

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Summary of the relevant parts of a struct/enum field.

Fields
span: Span

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

name: Option<Ident>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

None for tuple structs/normal enum variants, Some for normal structs/struct enum variants.

self_: P<Expr>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

The expression corresponding to this field of self (specifically, a reference to it).

other: Vec<P<Expr>>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

The expressions corresponding to references to this field in the other Self arguments.

attrs: &'a [Attribute]

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

The attributes on the field