Struct syntax_ext::deriving::generic::MethodDef

pub struct MethodDef<'a> {
    pub name: &'a str,
    pub generics: LifetimeBounds<'a>,
    pub explicit_self: Option<Option<PtrTy<'a>>>,
    pub args: Vec<Ty<'a>>,
    pub ret_ty: Ty<'a>,
    pub attributes: Vec<Attribute>,
    pub is_unsafe: bool,
    pub unify_fieldless_variants: bool,
    pub combine_substructure: RefCell<CombineSubstructureFunc<'a>>,
}

Unstable (rustc_private)

Fields
name: &'a str
Unstable (rustc_private)

name of the method

generics: LifetimeBounds<'a>
Unstable (rustc_private)

List of generics, e.g. R: rand::Rng

explicit_self: Option<Option<PtrTy<'a>>>
Unstable (rustc_private)

Whether there is a self argument (outer Option) i.e. whether this is a static function, and whether it is a pointer (inner Option)

args: Vec<Ty<'a>>
Unstable (rustc_private)

Arguments other than the self argument

ret_ty: Ty<'a>
Unstable (rustc_private)

Return type

attributes: Vec<Attribute>
Unstable (rustc_private)
is_unsafe: bool
Unstable (rustc_private)
unify_fieldless_variants: bool
Unstable (rustc_private)

Can we combine fieldless variants for enums into a single match arm?

combine_substructure: RefCell<CombineSubstructureFunc<'a>>
Unstable (rustc_private)