Struct syntax_ext::deriving::generic::Substructure

pub struct Substructure<'a> {
    pub type_ident: Ident,
    pub method_ident: Ident,
    pub self_args: &'a [P<Expr>],
    pub nonself_args: &'a [P<Expr>],
    pub fields: &'a SubstructureFields<'a>,
}

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

All the data about the data structure/method being derived upon.

Fields
type_ident: Ident

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

ident of self

method_ident: Ident

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

ident of the method

self_args: &'a [P<Expr>]

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

dereferenced access to any Self_ or Ptr(Self_, _) arguments

nonself_args: &'a [P<Expr>]

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

verbatim access to any other arguments

fields: &'a SubstructureFields<'a>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?