Enum syntax_ext::deriving::generic::ty::Ty

pub enum Ty<'a> {
    Self_,
    Ptr(Box<Ty<'a>>, PtrTy<'a>),
    Literal(Path<'a>),
    Tuple(Vec<Ty<'a>>),
}

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

A type. Supports pointers, Self, and literals

Variants
Self_

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Ptr(Box<Ty<'a>>, PtrTy<'a>)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

&/Box/ Ty

Literal(Path<'a>)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

mod::mod::Type<[lifetime], [Params...]>, including a plain type parameter, and things like i32

Tuple(Vec<Ty<'a>>)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

includes unit

Methods

impl<'a> Ty<'a>

pub fn to_ty(
&self,
cx: &ExtCtxt,
span: Span,
self_ty: Ident,
self_generics: &Generics
) -> P<Ty>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

pub fn to_path(
&self,
cx: &ExtCtxt,
span: Span,
self_ty: Ident,
self_generics: &Generics
) -> Path

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?