Struct syntax_ext::proc_macro_impl::AttrProcMacro

pub struct AttrProcMacro {
    pub inner: fn(_: TsShim, _: TsShim) -> TsShim,
}

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Fields
inner: fn(_: TsShim, _: TsShim) -> TsShim

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?