Function syntax_ext::proc_macro_registrar::modify

pub fn modify(
    sess: &ParseSess, 
    resolver: &mut Resolver, 
    krate: Crate, 
    is_proc_macro_crate: bool, 
    is_test_crate: bool, 
    num_crate_types: usize, 
    handler: &Handler
) -> Crate

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?